+++ git_improved/commands/new_branch.py
@@ -1,9 +1,0 @@
-
-def new_feature_branch_command():
-    print('new_feature_branch_command')
-
-def new_bugfix_branch_command():
-    print('new_bugfix_branch_command')
-
-def new_doc_branch_command():
-    print('new_doc_branch_command')