+++ git_improved/constants.py
@@ -1,7 +1,12 @@
 
-BRANCHES_PREFIXES = (
-    'bugfix', 'fix', 
-    'documentation', 'doc', 
-    'feature', 
-    'structure',
-    'devops')
+BRANCHES_PREFIXES = {
+    'bugfix': 'Bugfix',
+    'fix': 'Bugfix', 
+    'documentation': 'Documentation', 
+    'doc': 'Documentation',
+    'feature': 'Feature', 
+    'structure': 'Structure',
+    'devops': 'CI/CD',
+    'ci': 'CI/CD',
+    'cd': 'CI/CD'
+}